requirements.txt
Flask==3.0.0
pytesseract==0.3.10
Pillow==10.1.0
gunicorn==21.2.0
Werkzeug==3.0.1
